+++ packages/fcl/src/current-user/normalize/__vsn.js
@@ -1,19 +1,29 @@
 export const SERVICE_PRAGMA = {
-  "@type": "Service",
-  "@vsn": "1.0.0",
+  f_type: "Service",
+  f_vsn: "1.0.0",
 }
 
 export const IDENTITY_PRAGMA = {
-  "@type": "Identity",
-  "@vsn": "1.0.0",
+  f_type: "Identity",
+  f_vsn: "1.0.0",
 }
 
 export const PROVIDER_PRAGME = {
-  "@type": "Provider",
-  "@vsn": "1.0.0",
+  f_type: "Provider",
+  f_vsn: "1.0.0",
 }
 
 export const USER_PRAGMA = {
-  "@type": "USER",
-  "@vsn": "1.0.0",
+  f_type: "USER",
+  f_vsn: "1.0.0",
+}
+
+export const POLLING_RESPONSE_PRAGMA = {
+  f_type: "PollingResponse",
+  f_vsn: "1.0.0",
+}
+
+export const COMPOSITE_SIGNATURE_PRAGMA = {
+  f_type: "CompositeSignature",
+  f_vsn: "1.0.0",
 }